[FIX] stock_account: move revalaution to valuation test file
They both share the common setup and use the same mechanism.
The test_stockvaluation.py will be rework in the future since a lot
of tests are similar and probably useless. But until them we keep
them at the same place
